pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "user-management-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Kyrylo Alieksieiev",email = "newkirilal@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.6.1,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "aioboto3 (>=15.5.0,<16.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic-extra-types (>=2.10.0,<3.0.0)",
    "greenlet (>=3.3.1,<4.0.0)",
    "phonenumbers (>=9.0.22,<10.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "types-aioboto3[s3] (>=15.5.0,<16.0.0)",
    "aio-pika (>=9.5.8,<10.0.0)",
    "aiormq (>=6.9.2,<7.0.0)",
    "redis (>=7.1.0,<8.0.0)",
]


[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
pytest-cov = ">=7.0.0,<8.0.0"
httpx = ">=0.28.1,<0.29.0"
pytest-asyncio = ">=1.3.0,<2.0.0"
pytest-env = ">=1.3.2,<2.0.0"
testcontainers = ">=4.14.1,<5.0.0"
ruff = ">=0.14.14,<0.15.0"
black = ">=26.1.0,<27.0.0"
mypy = ">=1.19.1,<2.0.0"

[tool.pytest.ini_options]
addopts = "-v -s"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
env = [
    "PYTHONPATH=apps/user-management-service/src",
]

[tool.coverage.run]
omit = [
    "tests/*",
    "*/tests/*",
    "*_test.py",
    "test_*.py",
    "alembic/*"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py312']

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true